services/flexy-guard-admin/requirements.txt
Flask==1.0.2
Flask-BasicAuth==0.2.0
pymongo==3.7.2
python-dotenv==0.10.1
jsonschema==3.0.1